Remove redundant UserAccount class
